Android/#85 fix

+++ android-app/StoreIt/app/src/main/java/com/storeit/storeit/fragments/FileViewerFragment.java
@@ -1,67 +1,62 @@
 package com.storeit.storeit.fragments;
 
+import android.accessibilityservice.GestureDescription;
 import android.content.Context;
+import android.content.DialogInterface;
+import android.graphics.Color;
 import android.net.Uri;
 import android.os.Bundle;
+import android.support.design.widget.CoordinatorLayout;
+import android.support.design.widget.Snackbar;
 import android.support.v4.app.Fragment;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.util.Log;
-import android.view.ContextMenu;
 import android.view.GestureDetector;
 import android.view.LayoutInflater;
-import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.MotionEvent;
 import android.view.View;
 import android.view.ViewGroup;
-
+import android.widget.EditText;
+import android.widget.Toast;
+import com.storeit.storeit.R;
 import com.storeit.storeit.activities.MainActivity;
 import com.storeit.storeit.adapters.ExplorerAdapter;
+import com.storeit.storeit.protocol.StoreitFile;
 import com.storeit.storeit.services.SocketService;
 import com.storeit.storeit.utils.FilesManager;
-import com.storeit.storeit.R;
-import com.storeit.storeit.protocol.StoreitFile;
+
+import java.io.File;
 
 public class FileViewerFragment extends Fragment {
 
-    private static final String ARG_PARAM1 = "param1";
-    private static final String ARG_PARAM2 = "param2";
-
-    private String mParam1;
-    private String mParam2;
     private ExplorerAdapter adapter;
     private OnFragmentInteractionListener mListener;
     private RecyclerView explorersRecyclerView;
+
+    private boolean mMoving = false; // We are moving a file
+
+    public ExplorerAdapter getAdapter() {
+        return adapter;
+    }
 
     public FileViewerFragment() {
 
     }
 
     public static FileViewerFragment newInstance(String param1, String param2) {
-        FileViewerFragment fragment = new FileViewerFragment();
-        Bundle args = new Bundle();
-        args.putString(ARG_PARAM1, param1);
-        args.putString(ARG_PARAM2, param2);
-        fragment.setArguments(args);
-        return fragment;
+        return new FileViewerFragment();
     }
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if (getArguments() != null) {
-            /*
-            mParam1 = getArguments().getString(ARG_PARAM1);
-            mParam2 = getArguments().getString(ARG_PARAM2);
-            */
-
-            mParam1 = "param1";
-            mParam2 = "param2";
-
-        }
     }
+
+    CoordinatorLayout coordinatorLayout;
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
@@ -73,6 +68,8 @@ public class FileViewerFragment extends Fragment {
         explorersRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
 
         FilesManager manager = ((MainActivity) getActivity()).getFilesManager();
+
+        coordinatorLayout = (CoordinatorLayout)rootView.findViewById(R.id.coordinatorLayout);
 
         adapter = new ExplorerAdapter(manager, getContext());
         explorersRecyclerView.setAdapter(adapter);
@@ -136,27 +133,110 @@ public class FileViewerFragment extends Fragment {
 
     @Override
     public boolean onContextItemSelected(MenuItem item) {
+        if (mMoving)
+            return super.onContextItemSelected(item);
+
         int position = adapter.getPosition();
 
-        MainActivity activity  = (MainActivity)getActivity();
-        FilesManager manager = activity.getFilesManager();
-        StoreitFile file = adapter.getFileAt(position);
-        SocketService service = activity.getSocketService();
+        MainActivity activity = (MainActivity) getActivity();
+        final FilesManager manager = activity.getFilesManager();
+        final StoreitFile file = adapter.getFileAt(position);
+        final SocketService service = activity.getSocketService();
 
         switch (item.getItemId()) {
             case R.id.action_delete_file:
                 Log.v("FileViewerFragment", "Delete");
-                manager.removeFile(file);
+                manager.removeFile(file.getPath());
                 adapter.removeFile(position);
                 service.sendFDEL(file);
-
+                break;
+            case R.id.action_delete_file_disk:
+                if (!file.isDirectory()) {
+                    File localFile = new File(manager.getFolderPath() + File.separator + file.getIPFSHash());
+                    if (!localFile.delete()) {
+                        Toast.makeText(getContext(), "Error while deleting file", Toast.LENGTH_SHORT).show();
+                    }
+                }
                 break;
             case R.id.action_rename_file:
-                Log.v("FileVIewerFragment", "Rename");
+                renameFile(manager, file);
                 break;
+            case R.id.action_move_file:
+                View focus = getActivity().getCurrentFocus();
+                if (focus == null)
+                    break;
+
+                mMoving = true;
+                ((MainActivity)getActivity()).getFloatingButton().setVisibility(View.GONE); // floating button
+
+                Snackbar snackbar = Snackbar.make(focus, "Move file", Snackbar.LENGTH_INDEFINITE)
+                        .setAction("MOVE", new View.OnClickListener() {
+                            @Override
+                            public void onClick(View view) {
+                                ((MainActivity)getActivity()).getFloatingButton().setVisibility(View.VISIBLE);
+                                mMoving = false;
+
+                                String oldPath = file.getPath();
+                                String movedPath = getCurrentFile().getPath() +File.separator + file.getFileName();
+                                movedPath = movedPath.replace("//", "/");
+
+                                manager.removeFile(file.getPath());
+
+                                file.setPath(movedPath);
+
+                                manager.addFile(file, getCurrentFile());
+                                adapter.reloadFiles();
+                                service.sendFMOV(oldPath, movedPath);
+                            }
+                        });
+                snackbar.setCallback(new Snackbar.Callback() {
+                    @Override
+                    public void onDismissed(Snackbar snackbar, int event) {
+                        super.onDismissed(snackbar, event);
+                        ((MainActivity)getActivity()).getFloatingButton().setVisibility(View.VISIBLE);
+                        mMoving = false;
+                    }
+                });
+
+                snackbar.setActionTextColor(Color.RED);
+                snackbar.show();
+                break;
+
         }
 
         return super.onContextItemSelected(item);
+    }
+
+    private void renameFile(final FilesManager manager, final StoreitFile file) {
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+        LayoutInflater inflater = getActivity().getLayoutInflater();
+
+        builder.setTitle("Save picture");
+        View dialogView = inflater.inflate(R.layout.dialog_name_file, null);
+        builder.setView(dialogView);
+
+        final EditText input = (EditText) dialogView.findViewById(R.id.dialog_file_name_input);
+
+        builder.setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialogInterface, int i) {
+
+                String fileName = input.getText().toString();
+                File f = new File(file.getPath()); // Get parent path
+
+                String finalName = f.getParent() + File.separator + fileName;
+                finalName = finalName.replace("//", "/");
+                ((MainActivity) getActivity()).getSocketService().sendFMOV(file.getPath(), finalName);
+                manager.moveFile(file.getPath(), finalName);
+                adapter.reloadFiles();
+
+            }
+        }).setNegativeButton(android.R.string.cancel, new DialogInterface.OnClickListener() {
+            @Override
+            public void onClick(DialogInterface dialogInterface, int i) {
+                dialogInterface.cancel();
+            }
+        }).show();
     }
 
     public interface OnFragmentInteractionListener {
